+++ DataContext/Requests/OpenAccountRequest.cs
@@ -1,3 +1,3 @@
 namespace DataContext.Requests;
 
-public record struct OpenAccountRequest(long CustomerId, double InitialDeposit, int AccountTypeId);
+public record struct OpenAccountRequest(int CustomerId, double InitialDeposit, int AccountTypeId);
