+++ DataContext/Responses/DepositResponse.cs
@@ -2,7 +2,7 @@ using DataContext.Models;
 
 namespace DataContext.Responses;
 
-public record DepositResponse(long CustomerId, long AccountId, decimal Balance, bool Succeeded)
+public record DepositResponse(int CustomerId, int AccountId, decimal Balance, bool Succeeded)
     : IAccountResponse
 {
     public DepositResponse(IAccount account) : this(
